Рефакторинг добавления заказа

+++ RepositoryService/RepositoryService.Application/Models/ProductDTO.cs
@@ -8,7 +8,7 @@ namespace RepositoryService.Application.Models
         public string Article { get; init; }
         public required string Name { get; init; }
         public string? Description { get; init; }
-        public float Price { get; init; }
+        public decimal Price { get; init; }
         public int QuantityInStock { get; init; }
     }
 }
